requirements.txt
streamlit
pandas
pandasql
python-dotenv
groq
semantic-router[local]==0.0.72
chromadb
# sentence_transformers